requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
bcrypt==4.2.0
discord==2.3.2
discord.py==2.4.0
frozenlist==1.4.1
greenlet==3.1.0
idna==3.8
multidict==6.0.5
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.2
python-decouple==3.8
SQLAlchemy==2.0.34
telegram==0.0.1
typing_extensions==4.12.2
yarl==1.9.11
